Issue #142 move the compress_image function to trips/views.py
