+++ trips/models.py
@@ -2,9 +2,6 @@ from django.db import models
 from django.utils import timezone
 from members.models import Member
 from datetime import timedelta
-from PIL import Image
-from io import BytesIO
-
 
 class Trip(models.Model):
     name = models.CharField(max_length=100)
@@ -21,14 +18,6 @@ class Trip(models.Model):
             self.start_date + timedelta(days=x) 
             for x in range((self.end_date - self.start_date).days + 1)
         ]
-    
-    def compress_image(image):
-        img = Image.open(image)
-        img.thumbnail((200, 200))
-        output = BytesIO()
-        img.save(output, format="PNG", quality=70)
-        output.seek(0)
-        return output
 
 class TripMember(models.Model):
     trip = models.ForeignKey(Trip, on_delete=models.CASCADE)
